Simplify searching for a suitable conf by using match_string; I also renamed the header file to be more descriptive
