+++ ec_memory_configuration.h
